+++ src/components/IncomeInfo/index.jsx
@@ -2,11 +2,11 @@ import React from 'react';
 import { Container, IncomeTitle, IncomeValue } from './styles';
 
 function IncomeInfo(props) {
-  const { title, value, color } = props;
+  const { title, value, color, prefix, sufix } = props;
   return (
     <Container>
       <IncomeTitle>{title}</IncomeTitle>
-      <IncomeValue color={color}>{value}</IncomeValue>
+      <IncomeValue color={color}>{prefix} {value.toLocaleString()} {sufix}</IncomeValue>
     </Container>
   )
 }
