+++ src/components/InfoTag/index.jsx
@@ -1,16 +1,19 @@
 import React from 'react';
+import { Grid } from '@material-ui/core';
 import { Container, Title, Subtitle } from './styles';
 import Card from '../Card';
 
 function InfoTag(props) {
   const { title, subtitle, prefix, sufix } = props;
   return (
-    <Card>
-      <Container>
-        <Title>{title}</Title>
-        <Subtitle>{prefix} {subtitle?.toLocaleString()} {sufix}</Subtitle>
-      </Container>
-    </Card>
+    <Grid item xs={2}>
+      <Card>
+        <Container>
+          <Title>{title}</Title>
+          <Subtitle>{prefix} {subtitle?.toLocaleString()} {sufix}</Subtitle>
+        </Container>
+      </Card>
+    </Grid>
   )
 }
